Create CrashReporting Android Native Module
